requirements.txt
databento>=1.2.0
pandas>=2.0.0
numpy>=1.26.0
pytz>=2023.3
python-dotenv>=1.0.2
zstandard>=0.22.0
